users can deletew own accounts

+++ ASPNETMVCBlogDemo-master/Controllers/ApplicationUsersController.cs
@@ -9,20 +9,53 @@ using System.Web;
 using System.Web.Mvc;
 using MVCBlogDemo.Models;
 
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
+using System.Web;
+using System.Web.Mvc;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
+using Microsoft.Owin.Security;
+using MVCBlogDemo.Models;
+using Owin;
+
 namespace MVCBlogDemo.Controllers
 {
-    [Authorize(Roles="Admin")]
     public class ApplicationUsersController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
+        private ApplicationDbContext db;
+        public UserManager<ApplicationUser> UserManager { get; private set; }
+                private IAuthenticationManager AuthenticationManager
+        {
+            get
+            {
+                return HttpContext.GetOwinContext().Authentication;
+            }
+        }
+
+        public ApplicationUsersController()
+            : this(new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(new ApplicationDbContext())))
+        {
+        }
+
+        public ApplicationUsersController(UserManager<ApplicationUser> userManager)
+        {
+            UserManager = userManager;
+            db = new ApplicationDbContext();
+        }
 
         // GET: ApplicationUsers
+        [Authorize(Roles = "Admin")]
         public async Task<ActionResult> Index()
         {
             return View(await db.Users.ToListAsync());
         }
 
         // GET: ApplicationUsers/Details/5
+        [Authorize(Roles = "Admin")]
         public async Task<ActionResult> Details(string id)
         {
             if (id == null)
@@ -38,6 +71,7 @@ namespace MVCBlogDemo.Controllers
         }
 
         // GET: ApplicationUsers/Create
+        [Authorize(Roles = "Admin")]
         public ActionResult Create()
         {
             return View();
@@ -48,6 +82,7 @@ namespace MVCBlogDemo.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public async Task<ActionResult> Create([Bind(Include = "Id,UserName,Email,PasswordHash,SecurityStamp,IsConfirmed")] ApplicationUser applicationUser)
         {
             if (ModelState.IsValid)
@@ -61,6 +96,7 @@ namespace MVCBlogDemo.Controllers
         }
 
         // GET: ApplicationUsers/Edit/5
+        [Authorize(Roles = "Admin")]
         public async Task<ActionResult> Edit(string id)
         {
             if (id == null)
@@ -80,6 +116,7 @@ namespace MVCBlogDemo.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Admin")]
         public async Task<ActionResult> Edit([Bind(Include = "Id,UserName,Email,PasswordHash,SecurityStamp,IsConfirmed")] ApplicationUser applicationUser)
         {
             if (ModelState.IsValid)
@@ -92,6 +129,7 @@ namespace MVCBlogDemo.Controllers
         }
 
         // GET: ApplicationUsers/Delete/5
+        [Authorize(Roles="Admin")]
         public async Task<ActionResult> Delete(string id)
         {
             if (id == null)
@@ -107,13 +145,30 @@ namespace MVCBlogDemo.Controllers
         }
 
         // POST: ApplicationUsers/Delete/5
+        [Authorize]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> DeleteConfirmed(string id)
+        public async Task<ActionResult> DeleteConfirmed(string id, bool? notAdmin)
         {
             ApplicationUser applicationUser = db.Users.Where(u => u.Id == id).First();
+            if (applicationUser == null)
+            {
+                return HttpNotFound();
+            }
             db.Users.Remove(applicationUser);
+            var favourites = await db.Favourites
+                .Where(f => f.User.ApplicationUser.Id == id)
+                .ToListAsync();
+            foreach (Favourite fav in favourites)
+            {
+                db.Favourites.Remove(fav);
+            }
             await db.SaveChangesAsync();
+            if (notAdmin != null)
+            {
+                AuthenticationManager.SignOut();
+                return RedirectToAction("Index", "Home");
+            }
             return RedirectToAction("Index");
         }
